[FIX] stock_picking_group_by_partner_by_carrier: cache
Fix SO picking_ids cache issue when procurement groups are merged.
action_confirm was not called preventing to trigger auto orderpoints.
The m2m relation between sale.order and stock.picking is changed to
computed on the picking side. So reading from the SO won't trigger the
recompute. To workaround this, I force a flush of the computed field
when the value is modified to ensure the read from the SO will collect
the right data from DB.